Updates stlport's cmath to provide isfinite, isinf, isnan and signbit as members of the std namespace
This is required by the latest WebKit which uses cmath and expects these
functions to be in the std namespace. We could modify WebKit to expect these
functions to be in the global namespace (as other platforms have done, see
JavaScriptCore/wtf/MathExtras.h). However, in the simulator build, these
functions are in the std namespace, as they should be. So the WebKit
modification would have to be dependent on the Android build type. Instead, it's
better to fix this in stlport.

The fix used to achieve this is copied from cmath from GNU C++ 4.2.

Change-Id: Ifdcecbf256b44c837adad370894d385198fe4603