[MIPS] enable STL build for Android
Signed-off-by: Chris Dearman <chris@mips.com>

The conditional compilation of __Named_exception has been broken since
STLport release 5.1.0 in Nov 2005.  This class has been unconditionally
included for all non-Mips compilers on all platforms, which was not the
original intent.  Since then, changes elsewhere in _ios_base.h and _exception.h
have (unintentionally?) required __Named_exception in all uses, but this
went unnoticed until the Mips port of Android.

The current change has the same treatment of __Named_exception on all
platforms and compilers, including Mips Android gcc, and also the
SGI o32 Ugen compiler.  An identical (unexplained) change by Mips is
in the NDK copy of STLport.

The conditional compilation checks around definitions and use of
__Named_exception should someday be made identical, by the STLport designer.

The condition
   !(defined(_MIPS_SIM) && defined(_ABIO32) && (_MIPS_SIM == _ABIO32))
is not obvious.  This dates from v3.11 in July 1998 when this STL
library was only for SGI machines.  It was then a test of whether the
code was being compiled by Mips' original 32-bit-ABI Ugen compier,
or instead SGI's second-generation MIPSpro compiler.  The reason
for handling these compilers differently is not documented anywhere,
but was likely a workaround for some issue in one compiler.
These same flags are now defined by gcc compilers for Mips.

It is safe to now eliminate the ABIO32 check, because no active STLport
users have tried building with the Ugen compiler for many years
(otherwise they would have encountered build errors).

Change-Id: I38474a6b6616714fe3a611100adacf418cc3ae1d