Fix 64 bit builds.
stlport doesn't expose localeconv and setlocale
anyway, so there really isn't much point in exporting
lconv either.

This is not a source compatible change, but it certainly
isn't the first such change we've made. Besides, lconv is
a dummy structure anyway and there's nothing useful to be
gained from it.

Change-Id: Ifeeb929fc43bfc7a7c522cae18b9916027f04b82